Bugfix
The update contains a fix for a bug where some excited-to-excited transition dipole moments are read wrong. The transition dipole moments between the excited singlets (Sn-Sm) are read correctly. But if there are excited singlets and triplets in the calculation, then the TDMs for Tn-Tm are read from the Sn-Sm section. If there are more triplets than singlets, then some T-T TDMs are correct and some are wrong.
With the bugfix, now all excited-to-excited TDMs are correctly read.